fix json renderer with generators that yield no data

+++ connect/reports/renderers/json.py
@@ -21,14 +21,16 @@ class JSONRenderer(BaseRenderer):
         if tokens[-1] != 'json':
             output_file = f'{tokens[0]}.json'
         if inspect.isgenerator(data):
+            has_data = False
             with open(output_file, 'w') as f:
                 f.write('[')
                 for item in data:
+                    has_data = True
                     f.write(f'{json.dumps(item)}, ')
-
-            with open(output_file, 'rb+') as f:
-                f.seek(-2, os.SEEK_END)
-                f.truncate()
+            if has_data:
+                with open(output_file, 'rb+') as f:
+                    f.seek(-2, os.SEEK_END)
+                    f.truncate()
             with open(output_file, 'a') as f:
                 f.write(']')
         else:
